refactor: consolidate duplicated renderFavoritesQuickAccess
Extracts the function into js/core/favorites.js, imported by both
app.js and recipe-database.js. Eliminates sync-risk from duplication.

Closes #233
